feat(decimales): les montants passent de float à Decimal (§ BI.1, 2/2)
32 colonnes en NUMERIC exact, 7 restent en FLOAT. Le type suit la SOURCE
de la valeur : Decimal pour ce que l'utilisateur saisit ou importe d'un
relevé (montants, quantités, prix, PRU, taux, quotités, surface), float
pour ce qu'un fournisseur de données de marché produit (cours, prix
actuel, frais de gestion, poids de composition) — un flottant binaire dès
la source, où une Decimal n'ajouterait qu'une fausse exactitude.

Frontières explicites : comptable en Decimal (reconstruction, PRU,
plus-values, patrimoine net, emprunts, budget, salaire) ; analytique en
flottant avec conversion à l'entrée (XIRR, courbes d'évolution — cumul
exact, points flottants —, ratios, scores, rapport de période). La
valorisation quantité × cours est l'unique frontière marché → montant.

Migration 47651f844317 : FLOAT → NUMERIC(28, échelle), avec normalisation
des valeurs stockées par la règle d'arrondi de l'application, pas par le
ROUND() de SQLite qui arrondit la valeur binaire. Éprouvée en montée,
descente et remontée ; index conservés.

Contrat d'API inchangé (champ float Pydantic ← Decimal, nombre JSON) ;
export de données toujours en nombres JSON, exact.

Révélé en chemin :
- réimporter un fichier de transactions IDENTIQUE comptait une « mise à
  jour » par ligne (float 0.001 ≠ Decimal 0.0010000000) — corrigé dans
  _normalise_pour_comparaison ;
- try mal placé dans l'import de données, défaut latent antérieur ;
- écran Analyse en 500 (division par nb_mois = 3.0), vu par la seule
  suite de bout en bout : test unitaire ajouté, qui échoue sur l'ancien
  code.

Des tests tolérants deviennent exacts : realized_gain == Decimal("-0.99")
au lieu de approx(…, abs=1e-6), 0.1 + 0.05 == Decimal("0.15"), et le
scénario « au centime près » passe ses dix montants à l'égalité stricte.

1 456 tests verts ; bout en bout 81/81 ; balayage des 68 routes GET sur la
base de démonstration, aucune erreur serveur.

+++ backend/app/services/rapport_service.py
@@ -70,6 +70,8 @@ def _valeur_epargne_a_date(db: Session, holdings: list[Holding], date_dt: dateti
     (même bloc de construction, historique réel + ancrage sur le coût d'acquisition, § S.3,
     que la courbe combinée du Tableau de bord) évaluée à une seule date plutôt qu'en série
     complète."""
+    # Valeurs lues sur une série de courbe, donc en flottant (§ BI.1) — tout ce
+    # rapport calcule en flottant, les sommes comptables y entrant converties.
     total = 0.0
     par_type: dict[str, float] = {}
     for h in holdings:
@@ -96,7 +98,7 @@ def _interets_epargne_periode(holdings: list[Holding], date_debut_dt: datetime, 
         for h in holdings
         if h.type_actif in revenus_passifs_service.TYPES_LIVRETS_AVEC_TAUX and h.valeur_estimee and h.taux_pct
     )
-    return interets_annuels * jours_periode / 365
+    return float(interets_annuels * jours_periode / 365)
 
 
 def _versements_declares_periode(db: Session, holdings: list[Holding], date_debut_dt: datetime, date_fin_dt: datetime) -> float | None:
@@ -119,7 +121,7 @@ def _versements_declares_periode(db: Session, holdings: list[Holding], date_debu
     )
     if not points:
         return None
-    return sum(p.versement for p in points)
+    return float(sum(p.versement for p in points))
 
 
 def compute_rapport_epargne_periode(db: Session, date_debut: str, date_fin: str, user_id: int) -> dict:
@@ -193,7 +195,7 @@ def compute_rapport_periode(db: Session, date_debut: str, date_fin: str, user_id
     # depuis l'origine. `valeur_realisee_cumulee` (ventes + dividendes + intérêts +
     # autres revenus, cumulée) fait déjà partie de chaque point de
     # `compute_portfolio_history`.
-    montant_investi_periode = performance_service.montant_investi_periode(db, user_id, date_debut, date_fin)
+    montant_investi_periode = float(performance_service.montant_investi_periode(db, user_id, date_debut, date_fin))
     if points:
         valeur_debut_stricte = _champ_strict_a_ou_avant(points, date_debut, "valeur_portefeuille")
         valeur_fin_stricte = _champ_strict_a_ou_avant(points, date_fin, "valeur_portefeuille")
